Move alarm systems to interfaces
